// src/types/_scenario-menu-open.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type TheMenuToOpen =
  | 'file'
  | 'edit'
  | 'view'
  | 'run'
  | 'kernel'
  | 'settings'
  | 'help';

export interface MenuOpenScenarioOptions {
  menu: TheMenuToOpen;
  [k: string]: any;
}