// src/types/_scenario-sidebars.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type TheSidebarsToOpen = string[];

export interface SidebarsScenarioOptions {
  sidebars: TheSidebarsToOpen;
  [k: string]: any;
}